# Profile/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервис для управления профилями пользователей.",
        "title": "Profile Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/profile/api/v1",
    "paths": {
        "/categories": {
            "get": {
                "description": "Возвращает полный список категорий профессий",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profession-category"
                ],
                "summary": "Получить все категории",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profession_categories_models.CategoryDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новую категорию (только для админов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profession-category"
                ],
                "summary": "Создать категорию профессий",
                "parameters": [
                    {
                        "description": "Данные категории",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profession_categories_models.CategoryDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profession_categories_models.CategoryDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/categories/parent/{parentID}": {
            "get": {
                "description": "Возвращает список категорий, привязанных к конкретному родителю",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profession-category"
                ],
                "summary": "Получить подкатегории по ParentID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent ID",
                        "name": "parentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profession_categories_models.CategoryDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parentID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "description": "Возвращает одну категорию по её уникальному идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profession-category"
                ],
                "summary": "Получить категорию по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profession_categories_models.CategoryDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_YoungFlores_Case_Go_Profile_internal_profession_categories_models.CategoryDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}